+++ components/Documents/helpers.js
@@ -47,3 +47,28 @@ export function getDocumentDescriptionValidators() {
         ]
     };
 }
+
+export function mergeDocumentsWithChanged(documents, changedDocuments) {
+    changedDocuments.forEach((changedDoc) => {
+        switch (changedDoc.statusId) {
+            case 'deleted':
+            case 'replaced':
+            case 'archived':
+                for (let i = 0; i < documents.length; i++) {
+                    if (documents[i].documentId) {
+                        if (documents[i].documentId === changedDoc.documentId) {
+                            documents.splice(i, 1);
+                            break;
+                        }
+                    } else if (documents[i].documentUnapprovedId) {
+                        if (documents[i].documentUnapprovedId === changedDoc.documentUnapprovedId) {
+                            documents.splice(i, 1);
+                            break;
+                        }
+                    }
+                }
+                break;
+        }
+    });
+    return changedDocuments.concat(documents);
+}