+++ components/Grid/index.js
@@ -97,6 +97,7 @@ class Grid extends Component {
 
                     return (<Row
                       key={i}
+                      rowIndex={i}
                       data={rowData}
                       checked={checked}
                       linkableColumns={linkableColumns}